Changed parameters to double.

See merge request !35

+++ src/model/commands/ClearScreenCommand.java
@@ -10,7 +10,7 @@ public class ClearScreenCommand extends TurtleCommand {
 		super(numParameters, name);
 	}
 
-	public double execute(int[] parameters, Turtle myTurtle, Controller view) {
+	public double execute(double[] parameters, Turtle myTurtle, Controller view) {
 		view.clearDisplay();
 		view.moveTo(new Point(0, 0));
 		return super.distance(parameters, myTurtle);
